#11, Cucumber test fix, Selenium tests renewed. All old test work, Gepard is prepared for final removal of JUnit 3.x dependency.
